ai-converter/pyproject.toml
[project]
name = "sonics-onnx-converter"
version = "0.1.0"
description = "Convert AI music detection models to ONNX format"
requires-python = ">=3.12"
dependencies = [
    "sonics @ git+https://github.com/awsaf49/sonics.git",
    "fire>=0.7.1,<0.8.0",
    "torchlibrosa>=0.1.0,<0.2.0",
    "onnx>=1.20.0,<2.0.0",
    "onnxscript>=0.2.0",
    "transformers>=4.46.0",
]
